Added test mobs spawning randomly and fixed players being disconnected when hitting a despawned entity.

+++ src/org/enderstone/server/Main.java
@@ -349,8 +349,8 @@ public class Main implements Runnable {
 		return null;
 	}
 
-	private int latestKeepAlive = 0;
-	private int latestChunkUpdate = 0;
+	private long latestKeepAlive = 0;
+	private long latestChunkUpdate = 0;
 
 	private void serverTick(long tick) {
 		for (EnderPlayer ep : onlinePlayers) {
